RAG_API/requirements.txt
fastapi
uvicorn
chromadb
pydantic
python-dotenv
PyMuPDF
langchain
langchain-community
langchain_openai
langchain_chroma
python-multipart
